Fixed all warnings on FC9 with Boost 1.35, except for one that seems to be a problem either with Boost or GCC 4.3.

+++ src/cc/warp/vstring.h
@@ -151,7 +151,7 @@ public:
 
     bool operator==(str_data_t const & o) const
     {
-        return size() == o.size() &&
+        return size() == (size_t)o.size() &&
             std::equal(begin(), end(), o.begin());
     }
